refactor(platform): rename index.tsx to log-detail.tsx
