+++ turbo/apps/platform/src/views/logs-page/log-detail/log-detail.tsx
